requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
commonmark==0.9.1
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.5.0
django-on-heroku==1.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
